[Test framework] Moving provider classes to specific package
