+++ test-framework/examples/tests/src/test/java/org/keycloak/test/examples/RunOnServerTest.java
@@ -10,8 +10,8 @@ import org.keycloak.testframework.annotations.InjectRealm;
 import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.realm.ManagedRealm;
 import org.keycloak.testframework.remote.providers.runonserver.RunOnServerException;
-import org.keycloak.testframework.runonserver.InjectRunOnServer;
-import org.keycloak.testframework.runonserver.RunOnServerClient;
+import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
+import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
 import org.opentest4j.AssertionFailedError;
 
 @KeycloakIntegrationTest
