+++ test-framework/remote-providers/src/main/resources/META-INF/services/org.keycloak.services.resource.RealmResourceProviderFactory
@@ -1,2 +1,2 @@
-org.keycloak.testframework.remote.providers.TimeOffSetRealmResourceProviderFactory
+org.keycloak.testframework.remote.providers.timeoffset.TimeOffSetRealmResourceProviderFactory
 org.keycloak.testframework.remote.providers.runonserver.RunOnServerRealmResourceProviderFactory
