+++ test-framework/remote/src/main/java/org/keycloak/testframework/remote/runonserver/TestClassServerSupplier.java
@@ -1,4 +1,4 @@
-package org.keycloak.testframework.runonserver;
+package org.keycloak.testframework.remote.runonserver;
 
 import org.keycloak.testframework.injection.InstanceContext;
 import org.keycloak.testframework.injection.RequestedInstance;
