+++ test-framework/remote/src/test/java/org/keycloak/testframework/remote/runonserver/TestClassServerTest.java
@@ -1,4 +1,4 @@
-package org.keycloak.testframework.runonserver;
+package org.keycloak.testframework.remote.runonserver;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
