+++ tests/base/src/test/java/org/keycloak/tests/admin/authentication/AuthenticatorConfigTest.java
@@ -37,8 +37,8 @@ import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
 import org.keycloak.testframework.events.AdminEventAssertion;
 import org.keycloak.testframework.events.AdminEvents;
 import org.keycloak.testframework.realm.ManagedRealm;
-import org.keycloak.testframework.runonserver.InjectRunOnServer;
-import org.keycloak.testframework.runonserver.RunOnServerClient;
+import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
+import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
 import org.keycloak.testframework.util.ApiUtil;
 import org.keycloak.tests.utils.Assert;
 import org.keycloak.tests.utils.admin.AdminEventPaths;
